+++ src/pupil_labs/realtime_api/streaming/__init__.py
@@ -1,4 +1,9 @@
-from .base import RTSPData, RTSPRawStreamer, receive_raw_rtsp_data
+from .base import (
+    RTSPData,
+    RTSPRawStreamer,
+    SDPDataNotAvailableError,
+    receive_raw_rtsp_data,
+)
 from .gaze import GazeData, RTSPGazeStreamer, receive_gaze_data
 from .video import RTSPVideoFrameStreamer, VideoFrame, receive_video_frames
 
@@ -11,5 +16,6 @@ __all__ = [
     "RTSPGazeStreamer",
     "RTSPRawStreamer",
     "RTSPVideoFrameStreamer",
+    "SDPDataNotAvailableError",
     "VideoFrame",
 ]